refactor: settings module created
