+++ src/app/base/components/preloader/preloader.component.scss
@@ -27,7 +27,7 @@
   border-radius: 100%;
   background-color: $color-black;
   position: absolute;
-  border: 2px solid var(--color-white-alt);
+  border: 1px solid var(--color-white-alt);
   animation: preloader 3s infinite ease-in-out alternate;
 }
