+++ src/app/settings/components/delete-item/delete-item.component.scss
