+++ src/app/settings/components/favourites-list/favourites-list-media.scss
