+++ src/app/settings/components/favourites-list/favourites-list.component.scss
