+++ src/app/settings/components/phone-validation/phone-validation.component.scss
