+++ src/app/settings/components/user-info/user-info.component.scss
