+++ src/app/settings/pages/user-profile/user-profile-adaptive.scss
