+++ src/app/settings/pages/user-profile/user-profile.component.ts
@@ -3,9 +3,9 @@ Component, OnDestroy, OnInit, ChangeDetectorRef, AfterViewChecked
 } from '@angular/core';
 import { Subscription } from 'rxjs';
 import { AuthService } from '../../../core/service/auth.service';
-import { SettingsService } from '../../../core/service/settings.service';
-import { IFavourite } from '../../../core/models/favourite.model';
-import { IUserInfo } from '../../../core/models/userinfo.model';
+import { SettingsService } from '../../servise/settings.service';
+import { IFavourite } from '../../models/favourite.model';
+import { IUserInfo } from '../../models/userinfo.model';
 import { ModeService } from '../../../core/service/mode.service';
 
 @Component({
@@ -60,7 +60,7 @@ export class UserProfileComponent implements OnInit, OnDestroy, AfterViewChecked
     this.cdr.detectChanges();
   }
 
-  public getFavs(): void {
+  getFavs(): void {
     if (this.isUserAuthenticated) {
       const favsObservable = this.settingsService.getUserFavourites();
       favsObservable.subscribe((data) => {
@@ -72,7 +72,7 @@ export class UserProfileComponent implements OnInit, OnDestroy, AfterViewChecked
     }
   }
 
-  public getUserInfo(): void {
+  getUserInfo(): void {
     if (this.isUserAuthenticated) {
       const userObservable = this.settingsService.getUserInfoSettings();
       userObservable.subscribe((data) => {
@@ -81,13 +81,13 @@ export class UserProfileComponent implements OnInit, OnDestroy, AfterViewChecked
     }
   }
 
-  public userDataHandler(userData: IUserInfo): void {
+  userDataHandler(userData: IUserInfo): void {
     this.modelUser = userData;
     this.settingsService.SaveUserInfo(this.modelUser);
     this.cdr.detectChanges();
   }
 
-  public favouriteDeleteHandler(projectId: number): void {
+  favouriteDeleteHandler(projectId: number): void {
     this.settingsService.removeUserFavourite(projectId.toString()).subscribe((data) => {
       this.userFavourites = data.favourites;
       this.getFavs();
