+++ src/app/settings/settings.module.ts
@@ -1,0 +1,25 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { SharedModule } from '../shared/shared.module';
+import { UserInfoComponent } from './components/user-info/user-info.component';
+import { FavouritesListComponent } from './components/favourites-list/favourites-list.component';
+import { DeleteItemComponent } from './components/delete-item/delete-item.component';
+import { UserProfileComponent } from './pages/user-profile/user-profile.component';
+
+@NgModule({
+  declarations: [
+    UserProfileComponent,
+    UserInfoComponent,
+    FavouritesListComponent,
+    DeleteItemComponent
+  ],
+  imports: [
+    CommonModule,
+    SharedModule
+  ],
+  exports: [
+    UserProfileComponent
+  ]
+})
+export class SettingsModule { }
